added url for quick access to hosted game
I believe you can only have 1 github page for free at a time. Until I get another project to host and display, this link should stay valid. If the link is dead, it could be because I have hosted another page.

Please note again that this was originally developed on my minesweeper repo and I exported this code into this repo for hosting the content on github pages.